Update inference.py
During the inference phase when loading a trained model, there's a focus on exclusively loading the model's weights while disregarding or not loading the optimizer's state. This adjustment can be particularly beneficial when utilizing a pre-trained model for tasks such as inference or transfer learning, where the optimizer's state information might not be necessary.